moving original_dst cluster to extensions
