+++ source/common/upstream/cluster_manager_impl.cc
@@ -39,9 +39,9 @@
 #include "source/common/runtime/runtime_features.h"
 #include "source/common/tcp/conn_pool.h"
 #include "source/common/upstream/cds_api_impl.h"
+#include "source/common/upstream/cluster_factory_impl.h"
 #include "source/common/upstream/load_balancer_impl.h"
 #include "source/common/upstream/maglev_lb.h"
-#include "source/common/upstream/original_dst_cluster.h"
 #include "source/common/upstream/priority_conn_pool_map_impl.h"
 #include "source/common/upstream/ring_hash_lb.h"
 #include "source/common/upstream/subset_lb.h"
