+++ test/common/upstream/BUILD
@@ -72,6 +72,7 @@ envoy_cc_test(
         ":test_cluster_manager",
         "//source/common/router:context_lib",
         "//source/common/upstream:load_balancer_factory_base_lib",
+        "//source/extensions/clusters/original_dst:original_dst_cluster_lib",
         "//source/extensions/network/dns_resolver/cares:config",
         "//source/extensions/transport_sockets/tls:config",
         "//test/config:v2_link_hacks",
@@ -466,32 +467,6 @@ envoy_cc_test(
         "//test/test_common:utility_lib",
         "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
         "@envoy_api//envoy/config/core/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test(
-    name = "original_dst_cluster_test",
-    srcs = ["original_dst_cluster_test.cc"],
-    deps = [
-        ":utility_lib",
-        "//source/common/event:dispatcher_lib",
-        "//source/common/network:filter_state_dst_address_lib",
-        "//source/common/network:utility_lib",
-        "//source/common/upstream:original_dst_cluster_lib",
-        "//source/common/upstream:upstream_lib",
-        "//source/extensions/transport_sockets/raw_buffer:config",
-        "//test/mocks:common_lib",
-        "//test/mocks/local_info:local_info_mocks",
-        "//test/mocks/network:network_mocks",
-        "//test/mocks/protobuf:protobuf_mocks",
-        "//test/mocks/runtime:runtime_mocks",
-        "//test/mocks/server:admin_mocks",
-        "//test/mocks/server:instance_mocks",
-        "//test/mocks/ssl:ssl_mocks",
-        "//test/mocks/upstream:cluster_manager_mocks",
-        "//test/test_common:test_runtime_lib",
-        "//test/test_common:utility_lib",
-        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
     ],
 )
