+++ test/extensions/clusters/original_dst/BUILD
@@ -1,0 +1,35 @@
+load(
+    "//bazel:envoy_build_system.bzl",
+    "envoy_cc_test",
+    "envoy_package",
+)
+
+licenses(["notice"])  # Apache 2
+
+envoy_package()
+
+envoy_cc_test(
+    name = "original_dst_cluster_test",
+    srcs = ["original_dst_cluster_test.cc"],
+    deps = [
+        "//source/common/event:dispatcher_lib",
+        "//source/common/network:filter_state_dst_address_lib",
+        "//source/common/network:utility_lib",
+        "//source/common/upstream:upstream_lib",
+        "//source/extensions/clusters/original_dst:original_dst_cluster_lib",
+        "//source/extensions/transport_sockets/raw_buffer:config",
+        "//test/common/upstream:utility_lib",
+        "//test/mocks:common_lib",
+        "//test/mocks/local_info:local_info_mocks",
+        "//test/mocks/network:network_mocks",
+        "//test/mocks/protobuf:protobuf_mocks",
+        "//test/mocks/runtime:runtime_mocks",
+        "//test/mocks/server:admin_mocks",
+        "//test/mocks/server:instance_mocks",
+        "//test/mocks/ssl:ssl_mocks",
+        "//test/mocks/upstream:cluster_manager_mocks",
+        "//test/test_common:test_runtime_lib",
+        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/config/cluster/v3:pkg_cc_proto",
+    ],
+)
